Refactor session runtime into internal modules
